requirements.txt
python-dotenv
langchain
langchain_core
langchain_community
langchain_openai
streamlit